requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.4
